git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/branches/MPIDataMining@2412 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/basis/prewavelet/PrewaveletBasis.hpp
@@ -80,7 +80,7 @@ public:
 		else if (1 == index) //left border
 		{
 			//Index of the affected hatbasis. The affected bases are ab and ab + 1
-			int ab = floor(p * (1 << level));
+			int ab = static_cast<int>(floor(p * (1 << level)));
 			if (ab == 0)
 			{
 				return 0.9 * evalNormalHat(level, 1, p);
@@ -98,7 +98,7 @@ public:
 		else if ((unsigned int)(1 << level) - 1 == index) //right border
 		{
 			//Index of the affected hatbasis. The affected bases are ab and ab + 1
-			int ab = floor(p * (1 << level));
+			int ab = static_cast<int>(floor(p * (1 << level)));
 			if (ab == (1 << level) - 1)
 			{
 				return 0.9 * evalNormalHat(level, (1 << level) - 1, p);
@@ -117,7 +117,7 @@ public:
 		else
 		{
 			//Index of the affected hatbasis. The affected bases are ab and ab + 1
-			unsigned int ab = floor(p * (1 << level));
+			unsigned int ab = static_cast<int>(floor(p * (1 << level)));
 			if (ab == index - 3)
 			{
 				return normal_stamp[0] * evalNormalHat(level, ab + 1, p);
